set expacted type std::allocator_traits<svs::HugepageAllocator<float>>::difference_type to long for macos

+++ tests/svs/core/allocator.cpp
@@ -43,7 +43,7 @@ static_assert(std::is_same_v<typename Traits::const_pointer, const float*>);
 static_assert(std::is_same_v<typename Traits::void_pointer, void*>);
 static_assert(std::is_same_v<typename Traits::const_void_pointer, const void*>);
 #if defined(__APPLE__)
-static_assert(std::is_same_v<typename Traits::difference_type, int32_t>);
+static_assert(std::is_same_v<typename Traits::difference_type, long>);
 #else
 static_assert(std::is_same_v<typename Traits::difference_type, int64_t>);
 #endif // __APPLE__
